autoformat

+++ src/main/java/Game/TurretGenerator.java
@@ -3,7 +3,6 @@ package Game;
 import Game.Turrets.Turret;
 import Game.Turrets.Turret.Stats;
 import general.Log;
-import general.Util;
 import windowStuff.Button;
 import windowStuff.NoSprite;
 import windowStuff.Sprite;
@@ -15,16 +14,17 @@ public class TurretGenerator {
   private final MakeTurret func;
   private Turret pending = null;
 
-  public TurretGenerator(World world, String image, String label,MakeTurret make) {
-    this.world=world;
-    this.image=image;
-    this.label=label;
+  public TurretGenerator(World world, String image, String label, MakeTurret make) {
+    this.world = world;
+    this.image = image;
+    this.label = label;
     this.func = make;
   }
 
   public boolean generate(int x, int y) {
     Log.write("pad");
-    if (!world.canFitTurret(x, y, pending.stats[Stats.size]) || !world.tryPurchase(pending.stats[Stats.cost])) {
+    if (!world.canFitTurret(x, y, pending.stats[Stats.size]) || !world.tryPurchase(
+        pending.stats[Stats.cost])) {
       return false;
     }
 
@@ -38,13 +38,14 @@ public class TurretGenerator {
         new NoSprite(),
         this::generate);
     world.setCurrentTool(tool);
-    pending=func.make();
+    pending = func.make();
     tool.
         setOnMove(pending::move).
-        setOnDelete(()->{if(pending.isNotYetPlaced()) {
-          pending.delete();
-          this.pending=null;
-        }
+        setOnDelete(() -> {
+          if (pending.isNotYetPlaced()) {
+            pending.delete();
+            this.pending = null;
+          }
         });
   }
 
@@ -54,7 +55,7 @@ public class TurretGenerator {
           if (button == 0 && action == 1) {
             this.select();
           }
-        }, () -> label + ": cost= " + (int) 50);
+        }, () -> label + ": cost= " + 50);
   }
 
   @FunctionalInterface
